+++ server/api/embedded.go
@@ -1,0 +1,9 @@
+package api
+
+const (
+	EmbeddedPostPropAppID  = "appID"
+	EmbeddedPostPropDialog = "dialog"
+
+	EmbeddedResponseDataErrors = "errors"
+	EmbeddedResponseDataPost   = "post"
+)
